pyproject.toml
[project]
name = "brain-image-implementation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dreamsim",
    "torch",
    "torchvision",
    "torchsummary",
    "tensorboard",
    "torch_geometric",
    "lightning",
    "triton",

    "numpy",
    "einops",
    "pandas",
    "scikit-learn",
    "matplotlib",
    "seaborn",
    "Pillow",
    "jupyter",
    "ipykernel",

    "pytest",
    "tqdm",
    "pydantic",
    "hydra-core",
    "omegaconf",
]

[project.optional-dependencies]
test = ["pytest>=7.0.0", "pytest-cov>=4.0.0", "pytest-mock>=3.10.0"]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]


[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = ["--strict-markers", "--strict-config", "--verbose", "--tb=short"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]
